glibc: define __NR_copy_file_range.
danalbert noticed we spend 30s copying the 16GiB zip file of tests in
many builds, which made me look at cp performance, which made me realize
we're not using copy_file_range(2) in our toybox host prebuilt. Add the
missing #define to our host glibc sysroot to correct that. (This brings
the copy down from 30s to 20s, so the answer is still "don't do that",
but there's no point leaving performance on the table until we switch to
musl and bionic's up to date uapi headers.)

Test: rebuilt toybox
Change-Id: I99188a8c89ade34ffaa57492bc9b1cab5963e67c