add gcc tool inputs to filegroup
In order to sandbox rust compilation, all inputs need to be specified so that they can be copied to the sandbox. This CL topic adds explicit inputs and dependency relationships to the rust toolchain libraries so that they can be compiled in a sandbox, and so that other rust modules can properly specify their transitive dependencies on the toolchain.

Bug: 286077158
Change-Id: I12027aabc2966617723297d4b7927f3709595c79