string.h: add bionic's C23 memset_explicit() polyfill.
We're starting to get some users for this, and the lack of portability
is becoming a pain.

(Annoyingly both glibc and musl still only have explicit_bzero(), even
upstream. Which bionic _doesn't_ have. So although adding a polyfill for
_that_ to our ancient glibc might make more sense from the "what
glibc/musl upstream currently look like" perspective, it wouldn't
greatly improve our portability issue. Conversely, bionic could add
explicit_bzero() but it seems weird to be adding a non-standard
alternative to the standard function we already have.)

Change-Id: Idb2743a9c0909885aad96ad4ffaf3c62e59ec550